Extract all commands into its own module. This will allow for "Commands" to be extracted.
